PAS-51
- Change Login to Sign In
- Fix styling with Sign In page
- Modify TextButton to accept style overrides
- Modify text input sizing to be controlled by parent
