+++ .vscode/settings.json
@@ -1,5 +1,9 @@
 {
 	"editor.defaultFormatter": "esbenp.prettier-vscode",
 	"editor.formatOnSave": true,
-	"eslint.validate": ["typescript", "typescriptreact"]
+	"eslint.validate": [
+		"typescript",
+		"typescriptreact"
+	],
+	"typescript.tsdk": "node_modules/typescript/lib"
 }
