+++ src/pages/index.tsx
@@ -1,14 +1,14 @@
 import React from 'react';
 import { BrowserRouter as Router, Switch, Route } from 'react-router-dom';
 import GradeBook from '../GradeBook';
-import { Login, SignUp } from './accounts';
+import { SignIn, SignUp } from './accounts';
 
 function Pages() {
 	return (
 		<Router>
 			<Switch>
 				<Route path={`/sign-in`}>
-					<Login />
+					<SignIn />
 				</Route>
 				<Route path="/sign-up">
 					<SignUp />
